refactors code by moving subprocess management methods to cook/subprocess.py
